backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.100.0
uvicorn==0.22.0
python-multipart==0.0.20
pydantic==2.0.3
python-dotenv==1.0.0

# HTTP client and validation
requests==2.31.0
httpx==0.24.1
email-validator==2.0.0

# Image processing and validation
pillow==10.0.0

# Database (Supabase only - removed SQLAlchemy)
supabase==2.8.1
gotrue==2.8.1

# Configuration
pyyaml==6.0.1

# AI Services - Vision Agents
google-generativeai==0.3.0
openai==1.54.0
anthropic==0.40.0

# AI Services - Storyteller Agents  
mistralai==1.2.0
# openai and anthropic already listed above

# AI Services - Voice Agents
elevenlabs==1.14.0
google-cloud-texttospeech==2.14.1
azure-cognitiveservices-speech==1.40.0

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-mock==3.11.1
pytest-cov==4.1.0

# Development and linting
black==23.7.0
flake8==6.0.0
mypy==1.4.1
